Fix warnings
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@9291 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ apps/delaunayrefinement/Mesh.h
@@ -183,7 +183,7 @@ public:
         }
       }
     }
-    int msize = std::distance(mesh->active_begin(), mesh->active_end());
+    size_t msize = std::distance(mesh->active_begin(), mesh->active_end());
 
     if (found.size() != msize) {
       std::cerr << "Not all elements are reachable \n";
